Правит модуляцию
