+++ js/main.js
@@ -1,5 +1,3 @@
-import './util.js';
-import './get-cards.js';
-import './get-descriptions.js';
-import './get-likes.js';
-import './get-comments.js';
+import './utils.js';
+import './data.js';
+import './picture.js';
