+++ js/picture.js
@@ -1,0 +1,20 @@
+// На основе временных данных для разработки и шаблона #picture создайте DOM-элементы, соответствующие фотографиям, и заполните их данными:
+
+// Адрес изображения url подставьте как атрибут src изображения.
+// Количество лайков likes выведите в блок .picture__likes.
+// Количество комментариев comments выведите в блок .picture__comments.
+// Отрисуйте сгенерированные DOM-элементы в блок .pictures. Для вставки элементов используйте DocumentFragment.
+
+// Подключите модуль в проект.
+// Шаблон изображения случайного пользователя:
+//   <template id="picture">
+//     <a href="#" class="picture">
+//       <img class="picture__img" src="" width="182" height="182" alt="Случайная фотография">
+//       <p class="picture__info">
+//         <span class="picture__comments"></span>
+//         <span class="picture__likes"></span>
+//       </p>
+//     </a>
+//   </template>
+
+
